requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

